pyproject.toml
[project]
name = "data-mining-g6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.5.0",
    "numpy>=2.4.2",
    "scikit-learn>=1.8.0",
    "scipy-stubs~=1.17.0",
    "sentence-transformers>=5.2.3",
]
